Allow salary range to be floats in Jobs schema (#34)
* Resolve backend/main.py merge conflict from main branch

* Calculate UserStat's top_missing_skill, refractor ml/job-matches

* Remove unnecessary changes from repo

* Revert file to remote version

* Allow salary range to be floats in Jobs schema

+++ backend/models/job.py
@@ -6,9 +6,10 @@ from pydantic import BaseModel, Field, model_validator, ConfigDict, BeforeValida
 # This helper converts ObjectId to string for Pydantic validation
 PyObjectId = Annotated[str, BeforeValidator(str)]
 
+
 class SalaryRange(BaseModel):
-    min: Optional[int] = Field(None, ge=0)
-    max: Optional[int] = Field(None, ge=0)
+    min: Optional[float] = Field(None, ge=0)
+    max: Optional[float] = Field(None, ge=0)
     currency: Optional[str] = "USD"
 
     @model_validator(mode="after")
